Refactor Logger class
Acording to TEP8 the Logger functions managment and
the internal  attributes have to be moved to a new
intermediate class, using a Delegation pattern.

Implement LoggerHelper class to manage the Taurus Logger.
Most of the Logger method and attributes are moved to helper class,
and add a private attribute of LoggerHelper in Logger class.

Adapt the uses of the Logger in the Logger derivated class.